Updated test and error message to reflect the fact Windows _may_ write to the new log after moving source without a restart
