+++ src/Serilog.Sinks.EventLog/Sinks/EventLog/EventIdHash.cs
@@ -14,7 +14,7 @@
 
 using System;
 
-namespace Serilog.Formatting.Compact
+namespace Serilog.Sinks.EventLog
 {
     /// <summary>
     /// Hash functions for message templates. See <see cref="Compute"/>.
